+++ lib/components/components.dart
@@ -28,7 +28,7 @@ export 'swiper/indicator.dart';
 export 'swiper/pagination.dart';
 export 'swiper/swiper.dart';
 export 'switch.dart';
-export 'widget_decorator.dart';
+export 'decorator_box.dart';
 
 class NoScrollBehavior extends ScrollBehavior {
   @override